requirements/all.txt
amqp==5.1.1
arrow==1.2.2
asgiref==3.5.2
asttokens==2.0.5
attrs==21.4.0
Babel==2.10.1
backcall==0.2.0
billiard==3.6.4.0
blessed==1.19.1
celery==5.2.7
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coverage==6.4.2
cryptography==37.0.4
decorator==5.1.1
distlib==0.3.4
Django==4.1.1
django-appconf==1.0.5
django-cors-headers==3.13.0
django-cryptography==1.1
django-filter==22.1
django-jsonfield==1.4.1
django-money==2.1.1
django-phonenumber-field==6.1.0
django-picklefield==3.1
django-q==1.3.9
django-redis==5.2.0
django-sendfile==0.3.11
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
executing==0.8.3
filelock==3.7.1
flake8==4.0.1
flake8-docstrings==1.6.0
idna==3.3
iniconfig==1.1.1
ipython==8.4.0
jedi==0.18.1
kombu==5.2.4
matplotlib-inline==0.1.3
mccabe==0.6.1
model-bakery==1.5.0
packaging==21.3
parso==0.8.3
pexpect==4.8.0
phonenumbers==8.12.52
pickleshare==0.7.5
Pillow==9.2.0
platformdirs==2.5.2
pluggy==1.0.0
prompt-toolkit==3.0.29
psycopg2==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
py-moneyed==1.2
pycodestyle==2.8.0
pycparser==2.21
pydocstyle==6.1.1
pyflakes==2.4.0
Pygments==2.12.0
PyJWT==2.4.0
pyngrok==5.1.0
pyparsing==3.0.8
pytest==7.1.2
pytest-cov==3.0.0
pytest-django==4.5.2
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
redis==3.5.3
requests==2.28.1
six==1.16.0
snowballstemmer==2.2.0
sqlparse==0.4.2
stack-data==0.3.0
toml==0.10.2
tomli==2.0.1
tox==3.25.0
traitlets==5.3.0
urllib3==1.26.10
vine==5.0.0
virtualenv==20.14.1
wcwidth==0.2.5
